Core & Internals: rework *_session decorators. Closes #5947
Require 'session' to always be passed as a keyword argument to the
decorator and enforce this requirement by raising an exception.
Correctly update the signature of the returned function
("Session" -> "Optional[Session]"=None) and set the default value
of the 'session' kwarg to None.

There is currently no way to fully and correctly inform pyright about
this signature rewrite. That's why the type of decorators is not
entirely defined. I'm not sure how well the type deduction in pyright
will do its job, but I consider that it's more important to get rid
of the 500 + false positive errors in pyright related to session
management.