Core & Internals: enforce load before usage on RseData. Closes #5937
Explicitly raise an exception when trying to access an RseData
attribute which wasn't loaded first. This should avoid, in the future,
issues similar to the one being fixed.

Also, do a small optimisation which avoids doing a database call to
load the name of an RSE if this name was loaded before (via loading
rse_info or rse columns). This fixes the related issue.