Tests: test_rse_protocol_* to pytest
This ended being a much bigger rework than I would have liked.

The specialized protocol tests are now sub-classes of MgrTestCases,
this way the same tests are executed for each protocol without need
for copy-pasted boilerplate code. In order to de-activate a test
for a specific protocol, it's now sufficient to re-implement the
test function with an empty/skip logic in the subclass.
I had to disable the "dir deletion" test in ssh/rsync/rclone, but this
is not different from the previous behavior. This test was not run.

I removed the "static_file" logic which wasn't used in any tests.

I tried to reduce the amount of duplicate and copy-pasted code, but
I avoided code movement. Otherwise the commit would have been too big.
There is still a lot of copy-pasted code. To be factorized/improved
in another commit.

I removed the "mock" protocol test. Almost all tests fail for that
protocol, and I don't see reasons to even test it here.

I make use of the pytest builtin fixture "tmp_path_factory" to create
the temporary directory, so the cleanup is now automatic: by pytest.

The webdav/srm/gfal2 protocols where blindly refactored without
testing. However, these tests were already obviously broken for
a long time, as they used some python2 syntax.