Core & Internals: cleanup Distances table. Closes #5454
Most of the attributes are unused. The one which was actually used
in shortest path computations was called "ranking", which was confusing
for everybody, including myself.

Only keep the field which matters in the database. Rename it from
'ranking' to 'distance' to avoid confusion in the future. On CLI
side this means that the meaning of "distance" changes from
"agis_distance" to actual "distance". This seems more correct and
will probably be a welcomed change, but is an important behavioral
change to keep in mind.

Keep the possibility to set and get "ranking" as a synonym to
"distance". This is to maintain compatibility with older rucio clients.