Core & Internals: allow 1.30 to run with the 1.29 distances table
The correct column name is inferred automatically at startup, but
is not re-evaluated during run-time. A restart will be required after
updating the database schema.

Required to rework the client test initialization. This is to avoid
importing rse_factory in a context which doesn't have a database
configured. Otherwise the tests would fail because the rse factory
cannot be initialized correctly.