Tests: GFAL2 Tests Written; Fix #5962
Wrote the tests for the GFAL2 RSE Protocol, basing them on the tests
written for the WebDAV RSE Protocol. The tests use pytest.

In these tests, GFAL2 makes use of the underlying WebDAV protocol. A
connection is set up to a WebDAV server. The tests also make use of curl
to communicate with the server. The tests do not clean up the testing
environment after running.

Tests were not inherited from WebDAV (and are not coupled in any other
way) because the interfaces provided by the WebDAV and GFAL2 protocol
implementations are different. For example, the function signatures are
different in many methods. In addition, the WebDAV implementation will
apply the path2pfn method by itself in many of its methods, whereas GFAL
requires inputs to habe already passed through a path2pfn method.