Tests: SSL Certificates for WebDAV server; Fix #5962
Added lines to generate.sh to create the key pair for the WebDAV server
that the GFAL2 tests communicate with. GFAL2 uses the underlying WebDAV
protocol in the tests. In addition, GFAL2 requires the server it
communicates with to authenticate itself, hence the necessity for the
key pair.

Do not forget that you should not run the entire generate.sh file, only
portions of it (or else a new rucio CA will be generated).

The generated key pair has been baked into the WebDAV server image used
by the WebDAV and GFAL2 tests.

Also partially relevant for issue #6031 (testing the WebDAV RSE
Protocol).