Tests: Wrote WebDAV tests; Fix #6031
The WebDAV tests were completely rewritten to use pytest and make
maximum use of parametrisation and fixtures. Most methods of the WebDAV
protocol class are tested. Furthermore, timeout errors are specifically
tested for.

This implementation of tests does not make use of setup_ and teardown_class
methods to setup a testing folder before testing and delete the folder
after tests. This means that tests can be run concurrently, eg on github
with 3 runners. setup_method is called before each testcase and connects to
the WebDAV server with a fresh connection.

At the moment, there is no cleanup method after the tests.

Interactions with the server are also executed with calls to the command
line utility curl. This is done so that individual tests do not rely on
each other, and instead rely on curl to work.

Also added the WEB1 RSE (which is required for the tests) to the
activation script, including setting the distances to the other RSEs.