Tests: Fix Patch for Python 3.10; Fix #6025
test_db.py::test_pooloverload failed for python3.10, this has now been
fixed. The underlying issue is that when simply patching flask.request,
unittest will check whether flask.request is an async object. Running
this test without an application context causes flask to raise a
RuntimeError. The previous fix was hacky and did not work in python3.10.

The solution is to pass the patch command for flask.request a 'spec'
kwarg, this will prevent `_is_async_obj` (which checks if flask.request
is async) from being run. The class `T` is passed, which only implements
a class attribute called `method`. flask.request is patched by a mock
which only implements what is specified in class `T`.

The fact that the async probing by unittest causes flask to raise an
error is a known weakness of flask and unittest.