Authentication: Rework to raise CannotAuthenticate; Fix #5986
flaskapi/v1/common.py::request_auth_env makes a call to validate the
auth-token which it receives via the request header (as
"X-Rucio-Auth-Token"). In the call stack that is created to validate
this token, errors are propagated back to the flask environment by means
of "raise CannotAuthenticate" and "return None".

This confusing mixture means that in some cases, the more generic
"RucioException" is caught instead of the specific "raise
CannotAuthenticate", causing flask to return a HTTP500 (instead of
HTTP401, which would cause the client to attempt to reconnect).

This commit reworks the call stack to return a Dictionary for a
successfully validated token and raise CannotAuthenticate otherwise.
This streamlines the current process, which returns Optional[Dict] or
raises CannotAuthenticate.

The issue which triggered this patch was #5986, regarding the automatix
daemon.