Transfers: build all fts job params in one place. #6036
Before this commit, we generated job_params for each hop
of a path. After that, for multi-hop transfers, the params
of all hops were merged together in another part of code.
This made the code a bit difficult to follow and understand
what is the actual final job_params for a transfer.

Move all the code to the same function, which is now aware of
multihop transfers. This allows to directly set parameters
correctly. For example, bring_online always depend on the
first hop and is now directly generated from the transfer
definition of that hop rather than overwritten later in an
obscure way.

In the future, this will also allow to better handle multihop
parameters. For example correctly disable ipv6 when at least
one hop doesn't support it. Correctly handle s3 url styles
for source of a multihop. These changes are left for future
commits.

I removed the code related to source_spacetoken in job params,
because it was never set.