Transfers: compute path-wide checksum validation strategy. #6036
For multi-hop transfers, when a checksum validation strategy is set,
that strategy is used for each hop of the path. This can result in
transfer failures if the validation strategy is not permissive enough.
For example, assume a multi-hop:

RSE1(verify_checksum=False) -> RSE2(verify_checksum=True) -> RSE3(verify_checksum=True)

if we set the checksum strategy to 'both' by only checking the last
hop, the first hop transfer will fail, because RSE1 doesn't support
validating checksums. The strategy for the whole path must be the
less restrictive of all hops. In this case: "target", i.e. the checksum
will be validated only on RSE2 for first hop; and only on RSE3 for
second hop.

This commit ensures that we request a checksum validation strategy
compatible with all hops. This has a disadvantage: the checksum
may not be verified on hops which support it because some other hop
doesn't support validation. This is still better than failing
transfers though.