Core: add_rule with approval for file; Fix #5979
The issue was that core/rule.py::add_rule is designed to be applied on
DIDs with the boolean flag 'open'. When applied to files, the call to
the `did` dict with key 'open' is still performed, leading to a KeyError.

The KeyError is avoided by calling the dict value by using the `get`
method with a default value. This means that in the case the key-value
pair is not set, the value "Not Applicable" is returned (because we are
trying to find out whether a file and not a collection is open/closed).

In addition, the negation of the key is changed from `closed` to `open`
in the dict passed to template.safe_substitute. This simplifies the code
and follows the convention that Rucio tests whether an object is open
(and not whether it is closed). In accordance to this change, the
template file `rule_approval_request.tmpl` has been changed as well.

This issue did not show up in any tests so far because the code that
causes the error is only run when the --ask-approval flag is passed and
the file has rule_approvers assigned.

Wrote unit test to check for future errors in this workflow.