WebUI: create_did_sample uses JSON for new server endpoint; Fix #5608
In order to attach to the new server endpoint for create_did_sample,
the AJAX call within rucio.js::create_did_sample has been changed. The
parameters are now passed as JSON (instead of within the URL).

Also includes a script to populate the database so that one can manually
test the changes to the UI.