Client/Web: Strings passed via JSON; Fix #5608
In the didclient, information (such as scope and name of DIDs) is passed
through the URL. This has led to errors in
client/didclient.py::DIDClient::create_did_sample when attempting to
pass scope- and name-strings containing slashes. To circumvent this, the
data is now dumped into JSON and passed to the server in this form.

The server continues to accept requests in the old format, that has
remained unchanged.