auth: update x509 endpoint for browser clients #6048
The new webui requires the x509 endpoint to be updated to support
browser clients. This commit updates the endpoint to support both
browser clients and CLI clients.

The /auth/x509 endpoint can return HTTP 206 Partial Content to indicate
that multiple accounts were found for the given certificate and can
work with the new WebUI to pick an account.

The /auth/x509 endpoint can return more detailed error messages to the
webui ( idenity does not exist, identity not mapped to an account, ..).