Tests/Undertaker/Rule/DID Fixes for Undertaker; Fix #5154
In `code/did.py`, the process for DID-deletion was extended (both for
`temp_tables` and without). This was necessary because DIDs without
rules still require their replicas to be epoch-tombstoned (without the
`delete_rule`-process to handle that).

Took extra care not to add any SQL calls into the loop over all DIDs.

===

In `core/rule.py`, the case that no locks exist for a given rule was
handled. Since this would mean that the function
`__delete_lock_and_update_replica` (which sets the epoch tombstone)
would not be called, the epoch tombstone had to be set manually by
updating the Replica.

The Replica is updated by joining the ReplicationRule table to the
DataIdentifierAssociation table to the RSEFileAssociation table via the
scopenames. We select the RSEFileAssociation for which exists a rule
that indirectly protects it (rule assigned to dataset which is the
parent to a file which has a replica). We constrain this by asserting
that the rule needs to match a given rule_id, the replica must be
saved on an RSE matched by the rule's RSE Expression and the replica
must have a lock_cnt of 0.

The structure of the SQLAlchemy Query has been taken from other places
in the code. Choices such as using
`execution_options(synchronize_session=False)` stem from reusing
pre-existing code.