Transfers: set transfertool in submitter. Closes #5814
The submitter has some flexibility to submit with another transfertool
than the one initially requested. For example, when it fallbacks
to using another source than the one selected by preparer. In such
cases, it's important to correctly record the actual transfertool used
for submission, to be able to poll the transfer status from the correct
transfertool.