REST/API Change the `vos` api docs to OpenApi format; Fix #4992
Changed the obsolete API documentation for `vos.py`. The endpoints in
this file handle interactions with Virtual Organisations (creating them,
updating them, etc.). At the moment, this endpoint is only accessed by
`test_multi_vo.py` and no client exists for VOs. This might be an issue
for the future.