Tests/Judge Evaluator: Nested Containers
Added test `test_judge_double_container_with_existing_rule`. Sets up a
system of DIDs as follows:

`C1 (->) C2 -> D -> F`

Where `X->Y` means that `X` is the parent DID of `Y`. `C` is a
container, while `D` is a dataset and `F` is an associated file. `F` has
a replica initially located on RSE `R`. A rule is applied to `C1`
binding the container to `R`.

Initially, the connection in parentheses does not exist. The test first
attaches the structure `C2->D->F` to `C1` and runs the judge evaluator.
The expectation is that a datasetlock and a replicalock exist for `D`
and `F`, respectively. Each is expected to be in state `OK`.

The test then detaches the structure by detaching `C2` from `C1` and
runs the judge evaluator again. We expect there to be no dataset- or
replicalocks on `D` or `F`, respectively.

This test was written with regard to issue #5447 in the expectation of
triggering the bug described in said issue. This was not the case.