Tests/Judge Evaluator: Added double rule test; #5447
Added a test in the attempt to reproduce the bug described in issue #5447. In the end,
the bug could not be reproduced.

The test concerns the judge evaluator with two rules. The setup is as
follows:

    1. A container C is created and a first dataset D1 is added to C.
    2. A first rule A is created on the container C.
    3. Then another dataset D2 is added to the container C.
    4. Then a rule B is created on the container C.
    5. Run judge-evaluator.

We assert that both rules must be in state OK. Both rules must be
applied to each DID that can be traced back to the parent container C.