Multi VO: support certificate directory in conveyor. Close #6080
As implemented right now, rucio expects a config section with explicit
per-vo certificate paths:
```
[vo_certs]
vo1 = /path/to/cert1
vo2 = /path/to/cert2
```

While this works, it's quite a burden to configure multi-vo support.
Each time a new vo has to be added, multiple updates have to be done.
For example: the administrator has to explicitly update both the
parameters for the fts-renewal job, and add a new line in this list.

If rucio is allowed to automatically load certificates for 'vo's from
a directory, this reduces the number of places where things have to be
updated.