Testing: add a rucioclient container to docker files
Add a container which is identical to the rucio container,
but does nothing at all. It's just a placeholder.

This commit is a pre-requirement for usage of the intellij
docker-compose remote execution. When configuring intellij
remote interpreter, the user has to define the container in
which the commands will be executed. However, when you
actually execute the commands via the intellij UI, that
container is destroyed and re-created while overriding the
entry-point with intellij wrappers. As a result, we cannot
use the "rucio" container for that, because the httpd server
will be killed.

While I don't like adding such IDE-specific hacks upstream,
I don't see a better solution if we want to create an official
document about how to use intellij remote execution.
I maintained a local override patch since forever for that,
but we cannot require other developers to do the same thing.