WebUI: Handle CannotAuthenticate exception #6099
In #6045, the behaviour of validate_auth_token() in the core layer was
modified to raise a CannotAuthenticate exception in case of failure (as
opposed to returning None). The Web UI was not adapted to this change,
leading to the inability to authenticate using a brand-new session (or
one where the token has expired).