Transfers: apply hop penalty from the node, not towards. Fix #6170
This way it will not be applied on the first hop, while still correctly
using the RSE attribute for multi-hop transfers.

This is a slight behavioral change. Under some combinations of short
distances, more single hops will be preferred over multihops. But it's
unlikely that anybody will complain about this behavioral change.
Moreover, it's more in line with the definition of "hop_penalty".