Tests: fix sqlite tests with temp tabels by default. Closes #6145
A low number of tests use the db_session fixture. The goal is to
create a session and use it for all operations in the test. With the
activation of temporary tables by default in 1.31, some database
accesses, which where read-only before, now perform insertions into
temp tables. One example is list-replicas. In the case of sqlalchemy,
only one session is allowed to keep the write lock, so having two
sessions instantiated with a write access to the database will not
work.

Most of the tests using db_session where not correctly re-using the
same session for all operations. Moreover, the factories didn't
use at all db_session for initialization and cleanup. This resulted
in some cases when two write sessions tried to concurrently access
the sqlite database.

This commit re-works the rse and did factories to use db_session
for their database operations if that fixture is requested by the
calling test. This actually allowed to remove a dirty hack where
the db_session fixture required the rse_factory and did_factory
fixtures without obvious reasons (probably to trick pytest into
working as desired instead of fixing the underlying issue).

We also pass the db_session correctly to (hopefully) all functions
in tests which forgot to do it.

I'm not sure of the usefulness of re-using the same session object
throughout the whole test, but this is a different story. To be,
maybe, challenged, in another commit.

Also, consume list_replicas generators in test_replica_recoverer.
Otherwise the test tries to perform a REST call to server while
keeping a lock on the database.