Tests: implement non-global noparallel marks
It is now possible to have "noparallel groups". Tests with the same
group will never be run in parallel, but two different groups can be
run on two different nodes. If no groups are provided to
@pytest.mark.noparallel, the old behavior is maintained: that test
cannot be run concurrently at all.

The previous implementation was very convoluted:
We have implemented a pytest plugin, which had a hard dependency on
pytest-xdist. So we needed to ensure that these two plugins where
loaded in order. Our plugin also added some CLI arguments to
pytest: to allow choosing between native xdist and our overlay.

Another complication of the previous implementation came from the
fact that pytest-xdist only passes test names between the scheduler
node and the worker nodes. Our pytest.mark.noparallel is thus not
available in the scheduler node. To circumvent this limitation,
during the collection phase, the worker nodes where adding a "bogus"
non-existing test in front of the list of names sent to the scheduler.
The name of this test encoded (using json encoding) how many tests
where marked noparallel. Moreover, the noparallel tests where sorted
to be at the head end of the list sent to scheduler.

These two work-arounds, plus some other smaller ones, allowed to
achieve the desired outcome: tests marked no-parallel where executed
without parallelism. However, it was a dead-end in terms of how
intelligent we can make the scheduler.

The current commit completely re-works everything related to parallel
testing. We now rely on two hooks provided by the xdist plugin to
register or own logic into xdist. This includes two changes:
- we override the code on the worker (remote) node to send a dictionary
instead of a list to the scheduler node. This allows to tell the
scheduler any additional custom data indexed by the test name. In
particular, we send information about what tests are marked 'noparallel'
- we update the scheduler to treat the received data as a dict instead
of a list and change the scheduler logic to directly use the knowledge
about 'noparallel' marks during scheduling.

We plug too deeply to my liking into the pytest.xdist internals. Drastic
changes in pytest-xdist may break our code, but this was already the
case before this commit.

Also unchanged from the previous behavior: we still have to work around
the fact that pytest-xdist worker nodes only executes a test if they
have a second one (or a shutdown signal) in the queue.