Daemons: Add params and type hints to stop() Fixes #6075
The stop() functions used by daemons to gracefully stop had missing
parameters `signum` and `frame`. These parameters are passed
when the daemon receives SIGTERM signal. Since we have no use
for the parameter as of now, we have set them to `None` by default.
Also type-hints are added to the parameters and to the return type.

Signed-off-by: Soumik Dutta <shalearkane@gmail.com>