Core: Remove protocol reordering #6083
This functionality appears to be unwanted. For one thing, it is not
consistent (it applies for update and delete, but not for insert). For
another, it hinders the work of those configuring the RSEs by hand.

In the end, it should not matter for Rucio if there are gaps in the
priorities.