Core & Internals: finalize unified core/common config. Closes #6107
Avoid direct usage of core.config primitives. Don't set a default
argument to the convert_type_fnc. This should dissuade people from
using core.config.get directly, as it requires passing a mandatory
function attribute to convert the type.

For the API endpoints, keep the old behavior of automatically converting
the type to avoid breaking the existing API for clients.