Core: Do not restrict maximum protocol priority #6201
The motivation behind the removed piece of code appears to have been to
prevent gaps in the priorities of enabled protocols. However, this would
have been ineffective because 1) the same does not apply at the time the
protocol is added and 2) there could be more than one protocol with the
same priority.