Core & Internals: annotate sqlalchemy declarative models. #5578
Follow the steps from the following documentation:
https://docs.sqlalchemy.org/en/20/changelog/whatsnew_20.html#orm-declarative-models

Using the mapped_column and the Mapped annotation actually has an
important impact on sqlalchemy: columns which are not annotated to
return Optional[something] are automatically marked by sqlalchemy as
nullable = False. With the old, "Column", format, the default was
nullable = True (unless it's part of a Primary Key). As a consequence,
I had to manually go over every column decide if it should be annotated
with optional or not. I'm not 100% sure I correctly identified all of
them. Scream if you see something which doesn't seem right.