Core & Internals: don't use events to add deleted to SoftModel
SoftModel is a Model which is never deleted under normal rucio
operation. We have two of them: Identity and RSE. For some reason,
we use a multi-stage process to attach the 'deleted' and 'deleted_at'
columns to these models:
- When inheriting from SoftModel a 'soft_delete' mark is added to the
user-defined info dictionary of the table.
- An event is registered to add two columns to all tables having this
mark.

Instead of doing that, just implement directly the columns in the
SoftModel, the same way it's done in the Model for `created_at`
and `updated_at`.

This should get rid of many pyright errors related to missing the given
fields in the model.