Transfers: rework daemons to use a dispatcher queue. Closes #6190
Dissociate the code which gets the rows from the database from the
code doing the work on those rows. Communicate between the two using
a queue.

Some tests have to be slightly updated because they passed a `session`
argument to preparer and throttler. This is not possible anymore
because of using two separate threads for getting data from the
database and working on that data.