Daemons: rework logging of the daemon name
The log messages will not be automatically prefixed by the 'executable'
anymore. If it's desired to log it, `%(processName)s` must be added
to the `[common]logformat` configuration
(https://docs.python.org/3/library/logging.html#logrecord-attributes)
In case of logging to a json format, the data will be outputted to the
`{'process': {'name': ...}}` json path.

Also, don't print a hash prefix in messages. Because of it,
elastic cannot correctly identify identical messages and analysis
becomes more complex. If desired to filter on the messages from the
same thread, `%(thread)s` must be added to the `logformat`, allowing
to filter logs from a specific thread.

This has the advantage of correctly outputting the thread id and process
name even when logging.log is called directly instead of passing via the
decorated logger