Bin: Fix type error in argparse; Fix #6289
in `argparse.ArgumentParser`, the `type` argument allows for type
conversions, it is a function that is run on the input value. The issue
was that the input could be `string` or `None`, but the function
`float()` can only deal with `string` (or real numbers).

Fix: Changed the function to also handle `None` correctly.