Transfers: use normal update to set last_processed
The bulk sqlalchemy bulk update verifies that the number of updated
rows is equal to the number of parameters. This is problematic in
rucio. Requests could be removed by the judges or the servers for
example. This results in errors like the following one:
```
UPDATE statement on table 'requests' expected to update 959 row(s); 878 were matched.
```