Transfers: Do not skip requests failed during submission #6300
Following 21513abef, requests are now implicitly updated by
list_and_mark_transfer_requests_and_source_replicas() at the time they
are fetched for processing. However, there is a check in the Finisher to
silently skip recently-updated requests. As such, requests failed during
submission end up in an endless cycle which prevents their resubmission.

There does not seem to be a valid motivation for this. Requests in
states SUBMISSION_FAILED and LOST are not processed by any other daemon.
Requests in state SUBMITTING that are recently updated are already
skipped in _fetch_requests().

This commit removes that check.